SlurmGCP. Move TPU code into separate file + refactoring
